refactor(AbstractJsonConverter): util/converter 추상 클래스 common/converter로 이동
