+++ src/main/java/com/backend/allreva/common/converter/AbstractJsonListConverter.java
@@ -1,14 +1,14 @@
-package com.backend.allreva.util.converter;
+package com.backend.allreva.common.converter;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.persistence.AttributeConverter;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.ArrayList;
+import java.util.List;
 
-public abstract class AbstractJsonSetConverter<T> implements AttributeConverter<Set<T>, String> {
+public abstract class AbstractJsonListConverter<T> implements AttributeConverter<List<T>, String> {
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper()
             .setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY)
@@ -17,32 +17,32 @@ public abstract class AbstractJsonSetConverter<T> implements AttributeConverter<
 
     private final Class<T> elementType;
 
-    protected AbstractJsonSetConverter(final Class<T> elementType) {
+    protected AbstractJsonListConverter(final Class<T> elementType) {
         this.elementType = elementType;
     }
 
     @Override
-    public String convertToDatabaseColumn(final Set<T> set) {
-        if (set == null || set.isEmpty()) {
+    public String convertToDatabaseColumn(final List<T> list) {
+        if (list == null || list.isEmpty()) {
             return "[]";
         }
         try {
-            return OBJECT_MAPPER.writeValueAsString(set);
+            return OBJECT_MAPPER.writeValueAsString(list);
         } catch (JsonProcessingException e) {
-            throw new RuntimeException("Set을 JSON으로 변환하는 데 실패했습니다.", e);
+            throw new RuntimeException("List를 JSON으로 변환하는 데 실패했습니다.", e);
         }
     }
 
     @Override
-    public Set<T> convertToEntityAttribute(final String json) {
+    public List<T> convertToEntityAttribute(final String json) {
         if (json == null || json.isBlank()) {
-            return new HashSet<>();
+            return new ArrayList<>();
         }
         try {
             return OBJECT_MAPPER.readValue(
-                    json, OBJECT_MAPPER.getTypeFactory().constructCollectionType(Set.class, elementType));
+                    json, OBJECT_MAPPER.getTypeFactory().constructCollectionType(List.class, elementType));
         } catch (JsonProcessingException e) {
-            throw new RuntimeException("JSON을 Set으로 변환하는 데 실패했습니다.", e);
+            throw new RuntimeException("JSON을 List로 변환하는 데 실패했습니다.", e);
         }
     }
 }
