+++ src/main/java/com/backend/allreva/common/converter/ImageListConverter.java
@@ -1,7 +1,6 @@
 package com.backend.allreva.common.converter;
 
 import com.backend.allreva.common.model.Image;
-import com.backend.allreva.util.converter.AbstractJsonListConverter;
 import jakarta.persistence.Converter;
 
 @Converter
