+++ src/main/java/com/backend/allreva/module/concert/concert/infra/jpa/SellerSetConverter.java
@@ -1,7 +1,7 @@
 package com.backend.allreva.module.concert.concert.infra.jpa;
 
+import com.backend.allreva.common.converter.AbstractJsonSetConverter;
 import com.backend.allreva.module.concert.concert.domain.value.Seller;
-import com.backend.allreva.util.converter.AbstractJsonSetConverter;
 import jakarta.persistence.Converter;
 
 @Converter
